Extend sigreturn_prep_fpu_frame
The sigreturn_prep_fpu_frame was designed to recompute the fpu's data
pointer inside the build stack frame to match the address in the
remote stack (the stack of the restarted process).

Since the PowerPC support it appears that there is a need to fix
another address inside the built stack frame, and it is has been
addressed through the hack of the RT_SIGFRAME_FPU macro.

The support of the Transactional Memory state will require fixing more
addresses and the way the service is currently designed will not allow
that. This patch is making this service more generic by passing the
pointer of the complete remote signal frame instead of the FPU's data
area. The service is also renamed to remove the explicit FPU
reference.

Signed-off-by: Laurent Dufour <ldufour@linux.vnet.ibm.com>